[Refactor]: User 저장기능 분리
FirestoreSignUpManager에 User 저장기능 분리
